services/pricing_service/requirements.txt
aiohttp
redis
python-dotenv